Use Int instead of Int64 + more generic type handling

+++ src/ops/Diff.jl
@@ -12,11 +12,11 @@ Formula
 """
 mutable struct Diff{In}
     const buffer::Vector{In}
-    const lag::Int64
-    index::Int64
+    const lag::Int
+    index::Int
 
     Diff{In}(
-        lag::Int64=1
+        lag::Int=1
         ;
         init_value::In=zero(In)
     ) where {In} = new{In}(fill(init_value, lag), lag, 1)
